fix fatal error when two sessions run simultaneously
When two servers run simultaneously one may get a display port while the
other will complain that the display port was locked. Then the session
would do a variety of things (end immediately or hang). A loop was added
that checks if the session failed to start and restart it.