refactored ConnView, added more comments
The ConnView class is now inherits from Mustache. So it is truly a
Mustache View now. Any missing methods called to it are forwarded to its
session object as well as the scriptview object. This allows a lot of
freedom in the mustache templates when building a connection information
string.