added dynamic password change
If using `vncauth` to authenticate with VNC server a random password is
created and stored in a file to be used for authentication. Now when a
user connects to the server with that password, another random password
will be generated and used for authentication. This adds complexity on
the developer side to dynamically check for new passwords.